[process][debug][cosmetic] Move debug_show_task_list() to debug_utils.c
