+++ kernel/process.c
@@ -8,12 +8,9 @@
 #include <tilck/kernel/sched.h>
 #include <tilck/kernel/list.h>
 #include <tilck/kernel/kmalloc.h>
-#include <tilck/kernel/hal.h>
 #include <tilck/kernel/errno.h>
 #include <tilck/kernel/user.h>
 #include <tilck/kernel/debug_utils.h>
-#include <tilck/kernel/elf_utils.h>
-#include <tilck/kernel/term.h>
 
 #include <sys/prctl.h> // system header
 #include <sys/wait.h>  // system header
@@ -635,129 +632,4 @@ sptr sys_prctl(int option, uptr a2, uptr a3, uptr a4, uptr a5)
 
    printk("[TID: %d] Unknown option: %d\n", option);
    return -EINVAL;
-}
-
-static int debug_get_tn_for_tasklet_runner(task_info *ti)
-{
-   for (int i = 0; i < MAX_TASKLET_THREADS; i++)
-      if (get_tasklet_runner(i) == ti)
-         return i;
-
-   return -1;
-}
-
-enum task_dump_util_str {
-
-   HEADER,
-   ROW_FMT,
-   HLINE
-};
-
-static const char *
-debug_get_task_dump_util_str(enum task_dump_util_str t)
-{
-   static bool initialized;
-   static char fmt[80] = NO_PREFIX;
-   static char hfmt[80];
-   static char header[256];
-   static char hline_sep[256] = "+-----------+-------+-------+----------+";
-   static char *hline_sep_end = &hline_sep[sizeof(hline_sep)];
-
-   if (!initialized) {
-
-      int path_field_len = (term_get_cols(get_curr_term()) - 80) + 37;
-
-      snprintk(fmt + 4, sizeof(fmt) - 4,
-               "| %%-9d | %%-5d | %%-5d | %%-8s | %%-%ds |\n",
-               path_field_len);
-
-      snprintk(hfmt, sizeof(hfmt),
-               "| %%-9s | %%-5s | %%-5s | %%-8s | %%-%ds |\n",
-               path_field_len);
-
-      snprintk(header, sizeof(header), hfmt,
-               "tid", "pid", "ppid", "state", "path or kernel thread");
-
-      char *p = hline_sep + strlen(hline_sep);
-
-      for (int i = 0; i < path_field_len + 2 && p < hline_sep_end; i++, p++) {
-         *p = '-';
-      }
-
-      if (p < hline_sep_end)
-         *p++ = '+';
-
-      if (p < hline_sep_end)
-         *p++ = '\n';
-
-      initialized = true;
-   }
-
-   switch (t) {
-      case HEADER:
-         return header;
-
-      case ROW_FMT:
-         return fmt;
-
-      case HLINE:
-         return hline_sep;
-
-      default:
-         NOT_REACHED();
-   }
-}
-
-static int debug_per_task_cb(void *obj, void *arg)
-{
-   const char *fmt = debug_get_task_dump_util_str(ROW_FMT);
-   task_info *ti = obj;
-
-   if (!ti->tid)
-      return 0; /* skip the main kernel task */
-
-   const char *state = debug_get_state_name(ti->state);
-
-   if (!is_kernel_thread(ti)) {
-      printk(fmt, ti->tid, ti->pid,
-             ti->pi->parent_pid, state, ti->pi->filepath);
-      return 0;
-   }
-
-   char buf[128];
-   const char *kfunc = find_sym_at_addr((uptr)ti->what, NULL, NULL);
-
-   if (!is_tasklet_runner(ti)) {
-      snprintk(buf, sizeof(buf), "<kernel: %s>", kfunc);
-   } else {
-      snprintk(buf, sizeof(buf), "<kernel: %s[%d]>",
-               kfunc, debug_get_tn_for_tasklet_runner(ti));
-   }
-
-   printk(fmt, ti->tid, ti->pid, ti->pi->parent_pid, state, buf);
-   return 0;
-}
-
-static void debug_dump_task_table_hr(void)
-{
-   printk(NO_PREFIX "%s", debug_get_task_dump_util_str(HLINE));
-}
-
-void debug_show_task_list(void)
-{
-   printk(NO_PREFIX "\n\n");
-
-   debug_dump_task_table_hr();
-   printk(NO_PREFIX "%s", debug_get_task_dump_util_str(HEADER));
-
-   debug_dump_task_table_hr();
-
-   disable_preemption();
-   {
-      iterate_over_tasks(debug_per_task_cb, NULL);
-   }
-   enable_preemption();
-
-   debug_dump_task_table_hr();
-   printk(NO_PREFIX "\n");
 }
